+++ stream/start.go
@@ -19,7 +19,7 @@ import (
 //
 // 当 stream 为 nil 时，将根据直播流地址自动生成
 func StartAnchor(capturing *sync.Map, stream streamentity.IStream, anchor entity.Anchor, path string,
-	fileSizeThreshold int, handler hanlders.IHandler) error {
+	fileSizeThreshold int64, handler hanlders.IHandler) error {
 	// 此次是否是换新文件保存视频
 	// 用于当正在录播且isNewFile为真时，不退出
 	var isNewFile = false
@@ -106,7 +106,7 @@ LabelNewFile:
 // StartFlvAnchor 开始录制 flv 直播流
 //
 // 参数为 正在录制表、主播信息、临时文件存储路径（不需担心重名）、单视频大小、视频处理器
-func StartFlvAnchor(capturing *sync.Map, anchor entity.Anchor, path string, fileSizeThreshold int,
+func StartFlvAnchor(capturing *sync.Map, anchor entity.Anchor, path string, fileSizeThreshold int64,
 	handler hanlders.IHandler) error {
 	s := &flv.Stream{Stream: &streamentity.Stream{}}
 
@@ -118,7 +118,7 @@ func StartFlvAnchor(capturing *sync.Map, anchor entity.Anchor, path string, file
 // 参数为 正在录制表、主播信息、临时文件存储路径（不需担心重名）、单视频大小、视频处理器
 //
 // 下载m3u8视频（非直播）时，可下载到单个文件中，不能分文件保存，因为会重读m3u8文件，也就会重头开始下载
-func StartM3u8Anchor(capturing *sync.Map, anchor entity.Anchor, path string, fileSizeThreshold int,
+func StartM3u8Anchor(capturing *sync.Map, anchor entity.Anchor, path string, fileSizeThreshold int64,
 	handler hanlders.IHandler) error {
 	s := &m3u8.Stream{Stream: &streamentity.Stream{}}
